% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geno.R, R/Util.R
\name{checkIfSampleIDsExist}
\alias{checkIfSampleIDsExist}
\title{Check if sample identifiers are stored in a BGEN file (for BGEN v1.2)}
\usage{
checkIfSampleIDsExist(bgenFile)

checkIfSampleIDsExist(bgenFile)
}
\arguments{
\item{bgenFile}{a character of BGEN file. Sometimes, BGEN file does not include sample IDs. This information can be extracted from BGEN file. Please refer to https://www.well.ox.ac.uk/~gav/bgen_format/spec/v1.2.html for more details.}
}
\description{
Check if sample identifiers are stored in a BGEN file (for BGEN v1.2)

Check if sample identifiers are stored in a BGEN file (for BGEN v1.2)
}
\examples{

BGENFile = system.file("extdata", "example_bgen_1.2_8bits.bgen", package = "SAIGE")
checkIfSampleIDsExist(BGENFile)

BGENFile = system.file("extdata", "example_bgen_1.2_8bits.bgen", package = "SAIGE")
checkIfSampleIDsExist(BGENFile)
}
