% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAIGE_fitGLMM_fast.R
\name{fitNULLGLMM}
\alias{fitNULLGLMM}
\title{Fit the null logistic mixed model and estimate the variance ratio by a set of randomly selected variants}
\usage{
fitNULLGLMM(plinkFile = "", phenoFile = "", phenoCol = "",
  traitType = "binary", invNormalize = FALSE, covarColList = NULL,
  qCovarCol = NULL, sampleIDColinphenoFile = "", centerVariables = NULL,
  tol = 0.02, maxiter = 20, tolPCG = 1e-05, maxiterPCG = 500,
  nThreads = 1, Cutoff = 2, numMarkers = 30, skipModelFitting = FALSE,
  outputPrefix = "")
}
\arguments{
\item{plinkFile}{character. Path to plink file to be used for calculating elements of the genetic relationship matrix (GRM)}

\item{phenoFile}{character. Path to the phenotype file}

\item{phenoCol}{character. Column name for the trait e.g. "CAD"}

\item{traitType}{character. e.g. "binary" or "quantitative". By default, "binary"}

\item{invNormalize}{logical. Whether to perform the inverse normalization of the trait or not. E.g. TRUE or FALSE. By default, FALSE}

\item{covarColList}{vector of characters. Covariates to be used in the glm model e.g c("Sex", "Age")}

\item{qCovarCol}{vector of characters. Categorical covariates to be used in the glm model (NOT work yet)}

\item{sampleIDColinphenoFile}{character.  Column name for the sample IDs in the phenotype file e.g. "IID".}

\item{centerVariables}{vector of characters.  Covariates to be centered (around the mean) e.g. c("birthYear")}

\item{nThreads}{integer. Number of threads to be used. By default, 1}

\item{numMarkers}{integer (>0). Number of markers to be used for estimating the variance ratio. By default, 30}

\item{skipModelFitting}{logical.  Whether tp skip fitting the null model and only calculating the variance ratio, By default, FALSE. If TURE, the model file ".rda" is needed}

\item{outputPrefix}{character. Path to the output files with prefix.}
}
\value{
a file ended with .rda that contains the glmm model information, a file ended with .varianceRatio.txt that contains the variance ratio value, and a file ended with #markers.SPAOut.txt that contains the SPAGMMAT tests results for the markers used for estimating the variance ratio.
}
\description{
Fit the null logistic mixed model and estimate the variance ratio by a set of randomly selected variants
}
