% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geno.R, R/Util.R
\name{getSampleIDsFromBGEN}
\alias{getSampleIDsFromBGEN}
\title{Extract sample identifiers from BGEN file (for BGEN v1.2)}
\usage{
getSampleIDsFromBGEN(bgenFile)

getSampleIDsFromBGEN(bgenFile)
}
\arguments{
\item{bgenFile}{a character of BGEN file.}
}
\description{
Extract sample identifiers from BGEN file (for BGEN v1.2)

Extract sample identifiers from BGEN file (for BGEN v1.2)
}
\examples{

BGENFile = system.file("extdata", "example_bgen_1.2_8bits.bgen", package = "SAIGE")
getSampleIDsFromBGEN(BGENFile)

BGENFile = system.file("extdata", "example_bgen_1.2_8bits.bgen", package = "SAIGE")
getSampleIDsFromBGEN(BGENFile)
}
