% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAIGE_fitGLMM_fast.R
\name{fitNULLGLMM}
\alias{fitNULLGLMM}
\title{Fit the null logistic mixed model and estimate the variance ratio by a set of randomly selected variants}
\usage{
fitNULLGLMM(plinkFile = "", phenoFile = "", phenoCol = "",
  traitType = "binary", invNormalize = FALSE, covarColList = NULL,
  qCovarCol = NULL, sampleIDColinphenoFile = "", tol = 0.02,
  maxiter = 20, tolPCG = 1e-05, maxiterPCG = 500, nThreads = 1,
  Cutoff = 2, numMarkers = 30, skipModelFitting = FALSE,
  memoryChunk = 2, tauInit = c(0, 0), LOCO = FALSE, traceCVcutoff = 1,
  ratioCVcutoff = 1, outputPrefix = "")
}
\arguments{
\item{plinkFile}{character. Path to plink file to be used for calculating elements of the genetic relationship matrix (GRM)}

\item{phenoFile}{character. Path to the phenotype file}

\item{phenoCol}{character. Column name for the trait e.g. "CAD"}

\item{traitType}{character. e.g. "binary" or "quantitative". By default, "binary"}

\item{invNormalize}{logical. Whether to perform the inverse normalization of the trait or not. E.g. TRUE or FALSE. By default, FALSE}

\item{covarColList}{vector of characters. Covariates to be used in the glm model e.g c("Sex", "Age")}

\item{qCovarCol}{vector of characters. Categorical covariates to be used in the glm model (NOT work yet)}

\item{sampleIDColinphenoFile}{character.  Column name for the sample IDs in the phenotype file e.g. "IID".}

\item{nThreads}{integer. Number of threads to be used. By default, 1}

\item{numMarkers}{integer (>0). Number of markers to be used for estimating the variance ratio. By default, 30}

\item{skipModelFitting}{logical.  Whether to skip fitting the null model and only calculating the variance ratio, By default, FALSE. If TURE, the model file ".rda" is needed}

\item{memoryChunk}{integer or float. The size (Gb) for each memory chunk. By default, 4}

\item{tauInit}{vector of numbers. e.g. c(1,1), Unitial values for tau. For binary traits, the first element will be always be set to 1. If the tauInit is not specified, the second element will be 0.5 for binary traits.}

\item{LOCO}{logical. Whether to apply the leave-one-chromosome-out (LOCO) approach. By default, FALSE. This option has not been extensively tested.}

\item{traceCVcutoff}{float. The threshold for coefficient of variantion (CV) for the trace estimator to increase nrun. By default 1. suggested: 0.0025. This option has not been extensively tested.}

\item{ratioCVcutoff}{float. The threshold for coefficient of variantion (CV) for estimating the variance ratio. By default 1. suggested 0.001. This option has not been extensively tested.}

\item{outputPrefix}{character. Path to the output files with prefix.}
}
\value{
a file ended with .rda that contains the glmm model information, a file ended with .varianceRatio.txt that contains the variance ratio value, and a file ended with #markers.SPAOut.txt that contains the SPAGMMAT tests results for the markers used for estimating the variance ratio.
}
\description{
Fit the null logistic mixed model and estimate the variance ratio by a set of randomly selected variants
}
