% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAIGE_SPATest.R
\name{SPAGMMATtest}
\alias{SPAGMMATtest}
\title{Run single variant score tests with SPA based on the logistic mixed model.}
\usage{
SPAGMMATtest(dosageFile = "", dosageFileNrowSkip = 0,
  dosageFileNcolSkip = 0, dosageFilecolnamesSkip = c("SNPID", "CHR", "POS",
  "Allele0", "Allele1"), bgenFile = "", bgenFileIndex = "", vcfFile = "",
  vcfFileIndex = "", vcfField = "DS", sampleFile = "",
  idstoExcludeFile = "", idstoIncludeFile = "", rangestoExcludeFile = "",
  rangestoIncludeFile = "", chrom = "", start = 0, end = 0,
  minMAC = 1, minMAF = 0, minInfo = 0, phenoFile = "", phenoCol = "",
  traitType = "binary", invNormalize = FALSE, covarColList = NULL,
  qCovarCol = NULL, sampleIDColinphenoFile = "", centerVariables = NULL,
  GMMATmodelFile = "", varianceRatioFile = "", Cutoff = 2,
  SAIGEOutputFile = "", numLinesOutput = 1000, IsSparse = TRUE,
  IsOutputAFinCaseCtrl = FALSE)
}
\arguments{
\item{dosageFile}{character. Path to dosage file. Each line contains dosages for a marker to be tested}

\item{dosageFileNrowSkip}{integer(>=0). Number of lines to be skiped in the dosage file.}

\item{dosageFileNcolSkip}{integer(>=0). Number of columns to be skiped in the dosage file}

\item{dosageFilecolnamesSkip}{vector of characters. The column names of the skipped columns. Default: c("SNPID", "CHR", "POS", "Allele0", "Allele1")}

\item{bgenFile}{character. Path to bgen file}

\item{bgenFileIndex}{character. Path to the .bgi file (index of the bgen file)}

\item{vcfFile}{character. Path to vcf file}

\item{vcfFileIndex}{character. Path to index for vcf file by tabix, ".tbi" by "tabix -p vcf file.vcf.gz"}

\item{vcfField}{character. genotype field in vcf file to use. "DS" for dosages or "GT" for genotypes. By default, "DS".}

\item{sampleFile}{character. Path to the file that contains one column for IDs of samples in the dosage, vcf  or bgen file with NO header}

\item{idstoExcludeFile}{character. Path to the file containing variant ids to be excluded from the bgen or vcf file}

\item{idstoIncludeFile}{character. Path to the file containing variant ids to be included from the bgen or vcf file}

\item{rangestoExcludeFile}{character. Path to the file containing genome regions to be excluded from the bgen file. The file contains three columns for chromosome, start, and end respectively with no header}

\item{rangestoIncludeFile}{character. Path to the file containing genome regions to be included from the bgen file. The file contains three columns for chromosome, start, and end respectively with no header}

\item{chrom}{character. string for the chromosome to include from vcf file. If not speficied, the entire vcf will be tested}

\item{start}{numeric. start genome position to include from vcf file.}

\item{end}{numeric. end genome position to include from vcf file.}

\item{minMAC}{numeric. Minimum minor allele count of markers to test. By default, 1. The higher threshold between minMAC and minMAF will be used}

\item{minMAF}{numeric. Minimum minor allele frequency of markers to test. By default 0. The higher threshold between minMAC and minMAF will be used}

\item{minInfo}{numeric. Minimum imputation info of markers to test (in bgen file)}

\item{phenoFile}{character. Path to the phenotype file}

\item{phenoCol}{character. Column name for the trait e.g. "CAD"}

\item{traitType}{character. e.g. "binary" or "quantitative". By default, "binary"}

\item{invNormalize}{logical. Whether to perform the inverse normalization of the trait or not. E.g. TRUE or FALSE. iBy default, FALSE}

\item{covarColList}{vector of characters. Covariates to be used in the glm model e.g c("Sex", "Age")}

\item{qCovarCol}{vector of characters. Categorical covariates to be used in the glm model (NOT work yet)}

\item{sampleIDColinphenoFile}{character.  Column name for the sample IDs in the phenotype file e.g. "IID".}

\item{centerVariables}{vector of characters.  Covariates to be centered (around the mean) e.g. c("birthYear")}

\item{GMMATmodelFile}{character. Path to the input file containing the glmm model, which is output from previous step. Will be used by load()}

\item{varianceRatioFile}{character. Path to the input file containing the variance ratio, which is output from the previous step}

\item{Cutoff}{by default = 2 (SPA test would be used when p value < 0.05 under the normal approximation)}

\item{SAIGEOutputFile}{character. Path to the output file containing the SPAGMMAT test results}

\item{numLinesOutput}{numeric. Output results for how many marker each time.}

\item{IsSparse}{logical. Whether to exploit the sparsity of the genotype vector for less frequent variants to speed up the SPA tests or not for dichotomous traits. By default, TRUE}

\item{IsOutputAFinCaseCtrl}{logical. Whether to output allele frequency in cases and controls. By default, FALSE}
}
\value{
SAIGEOutputFile
}
\description{
Run single variant score tests with SPA based on the logistic mixed model.
}
