% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAIGE_SPATest.R
\name{SPAGMMATtest}
\alias{SPAGMMATtest}
\title{Run single variant score tests with SPA based on the logistic mixed model.}
\usage{
SPAGMMATtest(bgenFile = "", bgenFileIndex = "", vcfFile = "",
  vcfFileIndex = "", vcfField = "DS", savFile = "",
  savFileIndex = "", sampleFile = "", idstoExcludeFile = "",
  idstoIncludeFile = "", rangestoExcludeFile = "",
  rangestoIncludeFile = "", chrom = "", start = 1, end = 2.5e+08,
  IsDropMissingDosages = FALSE, minMAC = 0.5, minMAF = 0,
  maxMAFforGroupTest = 0.5, minInfo = 0, GMMATmodelFile = "",
  varianceRatioFile = "", SPAcutoff = 2, SAIGEOutputFile = "",
  numLinesOutput = 10000, IsSparse = TRUE,
  IsOutputAFinCaseCtrl = FALSE, IsOutputNinCaseCtrl = FALSE,
  LOCO = FALSE, condition = "", sparseSigmaFile = "",
  groupFile = "", kernel = "linear.weighted", method = "optimal.adj",
  weights.beta.rare = c(1, 25), weights.beta.common = c(0.5, 0.5),
  weightMAFcutoff = 0.01, r.corr = 0, IsSingleVarinGroupTest = TRUE,
  cateVarRatioMinMACVecExclude = c(0.5, 1.5, 2.5, 3.5, 4.5, 5.5, 10.5,
  20.5), cateVarRatioMaxMACVecInclude = c(1.5, 2.5, 3.5, 4.5, 5.5, 10.5,
  20.5), dosageZerodCutoff = 0.2,
  IsOutputPvalueNAinGroupTestforBinary = FALSE,
  IsAccountforCasecontrolImbalanceinGroupTest = TRUE)
}
\arguments{
\item{bgenFile}{character. Path to bgen file. Currently version 1.2 with 8 bit compression is supported}

\item{bgenFileIndex}{character. Path to the .bgi file (index of the bgen file)}

\item{vcfFile}{character. Path to vcf file}

\item{vcfFileIndex}{character. Path to index for vcf file by tabix, ".tbi" by "tabix -p vcf file.vcf.gz"}

\item{vcfField}{character. genotype field in vcf file to use. "DS" for dosages or "GT" for genotypes. By default, "DS".}

\item{savFile}{character. Path to sav file}

\item{savFileIndex}{character. Path to index for sav file .s1r}

\item{sampleFile}{character. Path to the file that contains one column for IDs of samples in the dosage, vcf, sav, or bgen file with NO header}

\item{idstoExcludeFile}{character. Path to the file containing variant ids to be excluded from the bgen file. The file does not have a header and each line is for a marker ID.}

\item{idstoIncludeFile}{character. Path to the file containing variant ids to be included from the bgen file. The file does not have a header and each line is for a marker ID.}

\item{rangestoExcludeFile}{character. Path to the file containing genome regions to be excluded from the bgen file. The file contains three columns for chromosome, start, and end respectively with no header}

\item{rangestoIncludeFile}{character. Path to the file containing genome regions to be included from the bgen file. The file contains three columns for chromosome, start, and end respectively with no header}

\item{chrom}{character. string for the chromosome to include from vcf file. Required for vcf file. Note: the string needs to exactly match the chromosome string in the vcf/sav file. For example, "1" does not match "chr1". If LOCO is specified, providing chrom will save computation cost}

\item{start}{numeric. start genome position to include from vcf file. By default, 1}

\item{end}{numeric. end genome position to include from vcf file. By default, 250000000}

\item{IsDropMissingDosages}{logical. whether to drop missing dosages (TRUE) or to mean impute missing dosages (FALSE). By default, FALSE. This option only works for bgen, vcf, and sav input.}

\item{minMAC}{numeric. Minimum minor allele count of markers to test. By default, 0.5. The higher threshold between minMAC and minMAF will be used}

\item{minMAF}{numeric. Minimum minor allele frequency of markers to test. By default 0. The higher threshold between minMAC and minMAF will be used}

\item{maxMAFforGroupTest}{numeric. Maximum minor allele frequency of markers to test in group test. By default 0.5.}

\item{minInfo}{numeric. Minimum imputation info of markers to test. By default, 0. This option only works for bgen, vcf, and sav input}

\item{GMMATmodelFile}{character. Path to the input file containing the glmm model, which is output from previous step. Will be used by load()}

\item{varianceRatioFile}{character. Path to the input file containing the variance ratio, which is output from the previous step}

\item{SPAcutoff}{by default = 2 (SPA test would be used when p value < 0.05 under the normal approximation)}

\item{SAIGEOutputFile}{character. Path to the output file containing the assoc test results}

\item{numLinesOutput}{numeric. Number of  markers to be output each time. By default, 10000}

\item{IsSparse}{logical. Whether to exploit the sparsity of the genotype vector for less frequent variants to speed up the SPA tests or not for dichotomous traits. By default, TRUE}

\item{IsOutputAFinCaseCtrl}{logical. Whether to output allele frequency in cases and controls. By default, FALSE}

\item{IsOutputNinCaseCtrl}{logical. Whether to output sample sizes in cases and controls. By default, FALSE}

\item{LOCO}{logical. Whether to apply the leave-one-chromosome-out option. By default, FALSE}

\item{condition}{character. For conditional analysis. Genetic marker ids (chr:pos_ref/alt if sav/vcf dosage input , marker id if bgen input) seperated by comma. e.g.chr3:101651171_C/T,chr3:101651186_G/A, Note that currently conditional analysis is only for bgen,vcf,sav input.}

\item{sparseSigmaFile}{character. Path to the file containing the sparseSigma from step 1. The suffix of this file is ".mtx".}

\item{groupFile}{character. Path to the file containing the group information for gene-based tests. Each line is for one gene/set of variants. The first element is for gene/set name. The rest of the line is for variant ids included in this gene/set. For vcf/sav, the genetic marker ids are in the format chr:pos_ref/alt. For bgen, the genetic marker ids should match the ids in the bgen file. Each element in the line is seperated by tab.}

\item{kernel}{character. For gene-based test. By default, "linear.weighted". More options can be seen in the SKAT library}

\item{method}{character. method for gene-based test p-values. By default, "optimal.adj". More options can be seen in the SKAT library}

\item{weights.beta.rare}{vector of numeric. parameters for the beta distribution to weight genetic markers with MAF <= weightMAFcutoff in gene-based tests.By default, "c(1,25)". More options can be seen in the SKAT library}

\item{weights.beta.common}{vector of numeric. parameters for the beta distribution to weight genetic markers with MAF > weightMAFcutoff in gene-based tests.By default, "c(0.5,0.5)". More options can be seen in the SKAT library}

\item{weightMAFcutoff}{numeric. Between 0 and 0.5. See document above for weights.beta.rare and weights.beta.common. By default, 0.01}

\item{r.corr}{numeric. bewteen 0 and 1. parameters for gene-based tests.  By default, 0.  More options can be seen in the SKAT library}

\item{IsSingleVarinGroupTest}{logical. Whether to perform single-variant assoc tests for genetic markers included in the gene-based tests. By default, FALSE}

\item{cateVarRatioMinMACVecExclude}{vector of float. Lower bound of MAC for MAC categories. The length equals to the number of MAC categories for variance ratio estimation. By default, c(0.5,1.5,2.5,3.5,4.5,5.5,10.5,20.5). If groupFile="", only one variance ratio corresponding to MAC >= 20 is used}

\item{cateVarRatioMaxMACVecInclude}{vector of float. Higher bound of MAC for MAC categories. The length equals to the number of MAC categories for variance ratio estimation minus 1. By default, c(1.5,2.5,3.5,4.5,5.5,10.5,20.5). If groupFile="", only one variance ratio corresponding to MAC >= 20 is used}

\item{dosageZerodCutoff}{numeric. In gene- or region-based tests, for each variants with MAC <= 10, dosages <= dosageZerodCutoff with be set to 0. By default, 0.2.}

\item{IsOutputPvalueNAinGroupTestforBinary}{logical. In gene- or region-based tests for binary traits. if IsOutputPvalueNAinGroupTestforBinary is TRUE, p-values without accounting for case-control imbalance will be output. By default, FALSE}

\item{IsAccountforCasecontrolImbalanceinGroupTest}{logical. In gene- or region-based tests for binary traits. If IsAccountforCasecontrolImbalanceinGroupTest is TRUE, p-values after accounting for case-control imbalance will be output. By default, TRUE}
}
\value{
SAIGEOutputFile
}
\description{
Run single variant score tests with SPA based on the logistic mixed model.
}
